RF: move tests to the module of the tested code
Golang tests should be next to the code they test, this allows the tests
to be run from the scope of module or with the _test module from the
scope of an external module.

This is also convention and expected by most go devs

Added a Makefile to ensure working commands and documented and run
