+++ Makefile
@@ -1,0 +1,7 @@
+run:
+	@go run main.go
+.PHONY: run
+
+test:
+	@go test ./...
+.PHONY: test
