+++ controllers/dns_server_test.go
@@ -1,39 +1,43 @@
-package tests
+package controllers_test
 
 import (
 	"net/http"
 	"net/http/httptest"
+	"strings"
 	"testing"
 	"web-check-go/controllers"
 
 	"github.com/gin-gonic/gin"
 )
 
-func TestFirewallHandler(t *testing.T) {
+func TestDnsServerHandler(t *testing.T) {
 	router := gin.Default()
-	firewallCtrl := &controllers.FirewallController{}
-	router.GET("/firewall", firewallCtrl.FirewallHandler)
+	dnsCtrl := &controllers.DnsServerController{}
+	router.GET("/dns", dnsCtrl.DnsServerHandler)
 
 	testCases := []struct {
 		name         string
 		url          string
 		expectedCode int
+		expectedBody string
 	}{
 		{
 			name:         "Missing URL",
 			url:          "",
 			expectedCode: http.StatusBadRequest,
+			expectedBody: `{"error":"url parameter is required"}`,
 		},
 		{
 			name:         "Valid URL",
-			url:          "example.com",
+			url:          "https://example.com",
 			expectedCode: http.StatusOK,
+			expectedBody: `{"domain":"example.com","dns":[{"address":"93.184.215.14","hostname":null,"dohDirectSupports":false}]}`,
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			req, err := http.NewRequest("GET", "/firewall?url="+tc.url, nil)
+			req, err := http.NewRequest("GET", "/dns?url="+tc.url, nil)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -44,6 +48,10 @@ func TestFirewallHandler(t *testing.T) {
 
 			if resp.Code != tc.expectedCode {
 				t.Errorf("Expected status code %d, got %d", tc.expectedCode, resp.Code)
+			}
+
+			if strings.TrimSpace(resp.Body.String()) != tc.expectedBody {
+				t.Errorf("Expected body '%s', got '%s'", tc.expectedBody, strings.TrimSpace(resp.Body.String()))
 			}
 		})
 	}
