+++ controllers/linked_pages_test.go
@@ -1,4 +1,4 @@
-package tests
+package controllers_test
 
 import (
 	"encoding/json"
@@ -11,14 +11,14 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestLegacyRankHandler(t *testing.T) {
+func TestGetLinksHandler(t *testing.T) {
 	router := gin.Default()
 
-	ctrl := &controllers.LegacyRankController{}
+	ctrl := &controllers.GetLinksController{}
 
-	router.GET("/legacy-rank", ctrl.LegacyRankHandler)
+	router.GET("/get-links", ctrl.GetLinksHandler)
 
-	req, err := http.NewRequest("GET", "/legacy-rank?url=www.google.com", nil)
+	req, err := http.NewRequest("GET", "/get-links?url=www.google.com", nil)
 	assert.NoError(t, err)
 
 	w := httptest.NewRecorder()
@@ -27,13 +27,11 @@ func TestLegacyRankHandler(t *testing.T) {
 
 	assert.Equal(t, http.StatusOK, w.Code)
 
-	var response controllers.RankResponse
+	var response controllers.LinkResponse
 	err = json.Unmarshal(w.Body.Bytes(), &response)
 	assert.NoError(t, err)
 
 	assert.NotNil(t, response)
-
-	assert.Equal(t, "www.google.com", response.Domain)
-
-	assert.True(t, response.IsFound || !response.IsFound)
+	assert.NotNil(t, response.Internal)
+	assert.NotNil(t, response.External)
 }
